# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APRly API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /optimizer/calculate:
    post:
      operationId: calculateOptimization
      tags: [optimizer]
      summary: Calculate debt interest savings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizerInput"
      responses:
        "200":
          description: Optimizer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizerResult"
  /optimizer/guest-lead:
    post:
      operationId: upsertGuestLead
      tags: [optimizer]
      summary: Persist pre-registration calculator cards as one guest debt lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertGuestLeadInput"
      responses:
        "200":
          description: Guest lead upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpsertGuestLeadResponse"
        "400":
          description: Validation error
...
  /auth/register-and-checkout:
    post:
      operationId: registerAndCheckout
      tags: [auth]
      summary: Validate signup, store pending registration, create Stripe Checkout (subscription + setup fee + trial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAndCheckoutInput"
      responses:
        "200":
          description: Open Checkout in a new tab using checkoutUrl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterAndCheckoutPayload"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorsResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorsResponse"
        "503":
          description: Stripe is not configured
  /auth/checkout/session-status:
    get:
      operationId: getCheckoutSessionStatus
      tags: [auth]
      summary: Poll checkout completion; issues session cookies when payment is confirmed and the user exists
      parameters:
        - name: stripeSessionId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Current checkout / auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionStatusResponse"
        "404":
          description: Unknown session
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Email and password login; sets HTTP-only cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Invalid credentials
...
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Rotate refresh token and re-issue access cookie
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Missing or invalid refresh token
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current user profile
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized

    patch:
      operationId: patchMe
      tags: [auth]
      summary: Update first and last name after checkout
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMeInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized

    delete:
      operationId: deleteMe
      tags: [auth]
      summary: Delete the authenticated user account
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Account deleted
        "401":
          description: Unauthorized
  /auth/me/password:
    patch:
      operationId: patchMePassword
      tags: [auth]
      summary: Change password for the authenticated user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMePasswordInput"
      responses:
        "204":
          description: Password updated
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorsResponse"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aprly_access
  schemas:
...
    OptimizerInput:
      type: object
      required: [totalDebt, interestRate]
      properties:
        totalDebt:
          type: number
          minimum: 0
        interestRate:
          type: number
          minimum: 0
          description: Current APR as a percent (e.g. 24.99)
        targetRate:
          type: number
          minimum: 0
          default: 2
          description: Target APR after optimization (default 2%)
        monthlyPayment:
          type: number
          minimum: 0
          nullable: true
    OptimizerResult:
      type: object
      required:
        - dailyInterestWaste
        - monthlyInterestWaste
        - monthlySavings
        - annualSavings
        - currentPayoffMonths
        - newPayoffMonths
        - newPayoffDate
        - currentRate
        - targetRate
      properties:
        dailyInterestWaste:
          type: number
        monthlyInterestWaste:
          type: number
        monthlySavings:
          type: number
        annualSavings:
          type: number
        currentPayoffMonths:
          type: integer
        newPayoffMonths:
          type: integer
        newPayoffDate:
          type: string
          format: date
        currentRate:
          type: number
        targetRate:
          type: number
...
    ImportCardItem:
      type: object
      required: [brand, balance, rate]
      properties:
        brand:
          type: string
          minLength: 1
        balance:
          type: number
          exclusiveMinimum: 0
        rate:
          type: number
          exclusiveMinimum: 0
          maximum: 100
        accountId:
          type: string
        source:
          type: string
          enum: [manual, plaid]
...
    PlanLeadStatus:
      type: string
      enum: [recommended, in_progress, won, denied]
...
    UpsertGuestLeadInput:
      type: object
      required: [guestSessionId, cards]
      properties:
        guestSessionId:
          type: string
          minLength: 8
          maxLength: 64
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        cards:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ImportCardItem"
    UpsertGuestLeadResponse:
      type: object
      required:
        - leadId
        - id
        - status
        - cardCount
        - totalBalance
        - totalEstimatedSavings
        - primaryBrand
        - currentApr
        - targetApr
        - createdAt
      properties:
        leadId:
          type: integer
        id:
          type: integer
        status:
          $ref: "#/components/schemas/PlanLeadStatus"
        cardCount:
          type: integer
        totalBalance:
          type: number
        totalEstimatedSavings:
          type: number
        primaryBrand:
          type: string
        currentApr:
          type: number
        targetApr:
          type: number
        createdAt:
          type: string
          format: date-time
...
    RegisterAndCheckoutInput:
      type: object
      required: [email, password, confirmPassword, termsAccepted]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 20
        confirmPassword:
          type: string
          minLength: 8
          maxLength: 20
        termsAccepted:
          type: boolean
        guestSessionId:
          type: string
          maxLength: 64
          description: Links pre-registration calculator debt lead to the new user
    RegisterAndCheckoutPayload:
      type: object
      required: [checkoutUrl, stripeSessionId]
      properties:
        checkoutUrl:
          type: string
          format: uri
        stripeSessionId:
          type: string
    CheckoutSessionStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, paid, failed, expired]
        user:
          nullable: true
          $ref: "#/components/schemas/AuthUserSummary"
    AuthUserSummary:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 200
    MeResponse:
      type: object
      required: [id, email, role, hasActiveSubscription]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
        hasActiveSubscription:
          type: boolean
          description: True when the user has an active Stripe subscription on file
    PatchMeInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 120
        lastName:
          type: string
          minLength: 1
          maxLength: 120
    PatchMePasswordInput:
      type: object
      required: [password, confirmPassword]
      properties:
        password:
          type: string
          minLength: 8
          maxLength: 20
        confirmPassword:
          type: string
          minLength: 8
          maxLength: 20
    FieldErrorsResponse:
      type: object
      required: [fieldErrors]
      properties:
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...