# lib/api-zod/src/generated/types/registerAndCheckoutInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APRly API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterAndCheckoutInput {
  /** @maxLength 254 */
  email: string;
  /**
   * @minLength 8
   * @maxLength 20
   */
  password: string;
  /**
   * @minLength 8
   * @maxLength 20
   */
  confirmPassword: string;
  termsAccepted: boolean;
  /**
   * Links pre-registration calculator debt lead to the new user
   * @maxLength 64
   */
  guestSessionId?: string;
}